Create pyproject.toml
This lists the dependencies for the bot. It definitely lists a lot of dependencies that aren't needed for the earthquake stuff, but right now I'm not able to prune it of unrelated stuff.
It's made to be parsed by the Poetry tool, but note that the version of Poetry that works with Python 2.7 (required by this project) has been completely deprecated.